add sblist_insert_sorted()
inserts into a presorted list using a binary search which calls
the user supplied comparefunc which matches the one from qsort
in calling convention and behaviour.

if only this function is used to insert into a (previously empty) list,
the list will automatically always be sorted as a result.
it can only be used on a list that was created by other means and then
sorted entirely once using sblist_sort().