option parser: support GNU-style long options like --foo too
admittedly the previous way to use long options only like -bar is
a bit uncommon.
maybe we should phase out the old way over time, so short flags
like -a -b can be combined to -ab too. but if we do it, that's far
away in the future, so users can gradually adapt.