add bmap, a binary-search map (similar to hashmap)
looks and feels like a hashmap, but behind the scenes a sorted key
array and binary search is used to find items.
since the design doesn't use buckets nor hashes, handling is much
simpler than a hashmap. usage is almost as simple as python's dict.
memory usage is also better, and there's far less code than typical
hashmap implementations, while offering very fast speed for smaller
lists with up to 1000 items, and reasonable speeds for upto 10000
items.

so unless dealing with Big Data (TM), this is a nice alternative.

a rewrite of this using the recently added tglist is considered, so
the API isn't written in stone yet.