tglist_free_values: implement in an aggregate-type-safe way
if the list value type is aggregate, like a struct, the array
index syntax bombs out. we need to manually calculate an offset
into the items memory, and dereference a pointer based on that
to produce valid C. there's a small optimization that the op
is only done when the type is pointer-sized; this should cause
it to get optimized away when not used, since the condition
result is known at compile time.